// src/utils/key.enum.ts
export enum Keys {
    ArrowUp = 'ArrowUp',
    ArrowDown = 'ArrowDown',
    ArrowLeft = 'ArrowLeft',
    ArrowRight = 'ArrowRight',
    Space = 'Space',
    Enter = 'Enter',
    Escape = 'Escape',
    Shift = 'Shift',
    Control = 'Control',
    Alt = 'Alt',
    Tab = 'Tab',
    CapsLock = 'CapsLock',
    Backspace = 'Backspace',
    Delete = 'Delete',
    A = 'KeyA',
    B = 'KeyB',
    C = 'KeyC',
    D = 'KeyD',
    E = 'KeyE',
    F = 'KeyF',
    G = 'KeyG',
    H = 'KeyH',
    I = 'KeyI',
    J = 'KeyJ',
    K = 'KeyK',
    L = 'KeyL',
    M = 'KeyM',
    N = 'KeyN',
    O = 'KeyO',
    P = 'KeyP',
    Q = 'KeyQ',
    R = 'KeyR',
    S = 'KeyS',
    T = 'KeyT',
    U = 'KeyU',
    V = 'KeyV',
    W = 'KeyW',
    X = 'KeyX',
    Y = 'KeyY',
    Z = 'KeyZ',
    Digit0 = 'Digit0',
    Digit1 = 'Digit1',
    Digit2 = 'Digit2',
    Digit3 = 'Digit3',
    Digit4 = 'Digit4',
    Digit5 = 'Digit5',
    Digit6 = 'Digit6',
    Digit7 = 'Digit7',
    Digit8 = 'Digit8',
    Digit9 = 'Digit9',
    F1 = 'F1',
    F2 = 'F2',
    F3 = 'F3',
    F4 = 'F4',
    F5 = 'F5',
    F6 = 'F6',
    F7 = 'F7',
    F8 = 'F8',
    F9 = 'F9',
    F10 = 'F10',
    F11 = 'F11',
    F12 = 'F12'
  }
  